Added files via upload
This is the solution of my ice rover control application. By this application data can be send to ice rover and also get response back from the rover if it stuck somewhere. When rover sends data if it stuck then android system will keep vibrating and beeping so that operator can notify.